+++ lib/components/Toolbar.js
@@ -3,7 +3,7 @@ import React from 'react';
 
 import DraftUtils from '../api/DraftUtils';
 
-import Button from '../components/Button';
+import ToolbarButton from '../components/ToolbarButton';
 import ToolbarGroup from '../components/ToolbarGroup';
 
 import { BR_TYPE, ENTITY_TYPE, UNDO_TYPE, REDO_TYPE } from '../api/constants';
@@ -39,7 +39,7 @@ const Toolbar = ({
     <div className="editor__toolbar" role="toolbar">
         <ToolbarGroup>
             {inlineStyles.map(style => (
-                <Button
+                <ToolbarButton
                     key={style.type}
                     name={style.type}
                     active={DraftUtils.hasCurrentInlineStyle(
@@ -56,7 +56,7 @@ const Toolbar = ({
 
         <ToolbarGroup>
             {blockTypes.map(block => (
-                <Button
+                <ToolbarButton
                     key={block.type}
                     name={block.type}
                     active={DraftUtils.isSelectedBlockType(
@@ -73,7 +73,7 @@ const Toolbar = ({
 
         <ToolbarGroup>
             {enableHorizontalRule ? (
-                <Button
+                <ToolbarButton
                     name={ENTITY_TYPE.HORIZONTAL_RULE}
                     onClick={addHR}
                     label="―"
@@ -82,7 +82,7 @@ const Toolbar = ({
             ) : null}
 
             {enableLineBreak ? (
-                <Button
+                <ToolbarButton
                     name={BR_TYPE}
                     onClick={addBR}
                     label="↵"
@@ -93,7 +93,7 @@ const Toolbar = ({
 
         <ToolbarGroup>
             {entityTypes.map(entity => (
-                <Button
+                <ToolbarButton
                     key={entity.type}
                     name={entity.type}
                     onClick={onRequestDialog.bind(null, entity.type)}
@@ -106,7 +106,7 @@ const Toolbar = ({
 
         <ToolbarGroup>
             {showUndoRedoControls ? (
-                <Button
+                <ToolbarButton
                     name={UNDO_TYPE}
                     onClick={onUndoRedo.bind(null, UNDO_TYPE)}
                     label="↺"
@@ -115,7 +115,7 @@ const Toolbar = ({
             ) : null}
 
             {showUndoRedoControls ? (
-                <Button
+                <ToolbarButton
                     name={REDO_TYPE}
                     onClick={onUndoRedo.bind(null, REDO_TYPE)}
                     label="↻"
