+++ lib/components/ToolbarButton.js
@@ -7,7 +7,7 @@ import Icon from '../components/Icon';
  * Displays a basic button, with optional active variant,
  * enriched with a tooltip. The tooltip stops showing on click.
  */
-class Button extends Component {
+class ToolbarButton extends Component {
     constructor(props) {
         super(props);
 
@@ -60,7 +60,7 @@ class Button extends Component {
     }
 }
 
-Button.propTypes = {
+ToolbarButton.propTypes = {
     name: PropTypes.string,
     active: PropTypes.bool,
     label: PropTypes.string,
@@ -69,7 +69,7 @@ Button.propTypes = {
     onClick: PropTypes.func,
 };
 
-Button.defaultProps = {
+ToolbarButton.defaultProps = {
     name: null,
     active: false,
     label: null,
@@ -78,4 +78,4 @@ Button.defaultProps = {
     onClick: () => {},
 };
 
-export default Button;
+export default ToolbarButton;
